Create Furniture Store Sales
Which country should this furniture store focus on marketing towards? 
Based on my analysis, this company should focus focus on marketing towards buyers in the United States. 
Current accounts in the United States make up the highest amount of sales per country and the highest amount of transactions per country. 
Visa and Mastercard are the two most used payment types for accounts in the United States. Further analysis could be done to determine if creating a promotion using these payment types could be beneficial for the furniture store.